driver: soc: ftm_alarm: forced conversion to u32 to fix a build warning
there is an building warning on 64bit platform nxp-ls1046 as below:
warning: large integer implicitly truncated to unsigned type [-Woverflow]
  ftm_writel(~0UL, ftm1_base + FTM_MOD);

convert to u32 to fix a build warning.

+++ drivers/soc/fsl/layerscape/ftm_alarm.c
@@ -135,7 +135,7 @@ static void ftm_clean_alarm(void)
 	ftm_counter_disable(ftm1_base);
 
 	ftm_writel(0x00, ftm1_base + FTM_CNTIN);
-	ftm_writel(~0UL, ftm1_base + FTM_MOD);
+	ftm_writel((u32)(~0UL), ftm1_base + FTM_MOD);
 
 	ftm_reset_counter(ftm1_base);
 }
